Rollback wro4j+less implementation.
See a solution when the project will be ready for prod. This is too early now for wro4j.
Less needs a modification in the server (to consider .less as a text/css mimetype).
I don't want any specific configuration, I'd like everyone to be able to install and run the project easily.